requirement.txt
fastapi
uvicorn

sqlalchemy
gunicorn

python-jose
passlib[bcrypt]
bcrypt

python-dotenv

google-generativeai

sentence-transformers
chromadb

beautifulsoup4
requests

streamlit
pydantic[email]